auto code-gen for creating ir operation in drr
