+++ paddle/fluid/pir/drr/drr_rewrite_pattern.h
@@ -22,7 +22,8 @@
 #include "paddle/fluid/pir/drr/api/drr_pattern_context.h"
 #include "paddle/fluid/pir/drr/api/match_context.h"
 #include "paddle/fluid/pir/drr/ir_operation.h"
-#include "paddle/fluid/pir/drr/ir_operation_creator.h"
+// #include "paddle/fluid/pir/drr/ir_operation_creator.h"
+#include "paddle/fluid/pir/drr/ir_operation_factory.h"
 #include "paddle/fluid/pir/drr/match_context_impl.h"
 #include "paddle/fluid/pir/drr/pattern_graph.h"
 #include "paddle/pir/core/enforce.h"
